Small redefines structure
class GamePlay extends Platformer
