--- src/libs/multiplatform/mobile/player-controls.ts
@@ -1,12 +1,0 @@
-import Resources from "../../class/resources";
-
-class PlayerControls {
-
-  private assets: Resources = new Resources();
-
-  constructor() {
-    // test
-  }
-
-}
-export default PlayerControls;
